requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Pillow==9.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dotenv==0.19.2
python-slugify==5.0.2
pytz==2021.3
slugify==0.0.1
sqlparse==0.4.2
text-unidecode==1.3
